removed performance model section
after some thinking and discussion with Melissa I have come to the conclusion that a performance model as initially envisioned (bigjob/script overhead) is currently not achievable since none of the other variables are fixed. for example, we don't know yet if a system with 64 12-core replicas has the same theoretical 'throughput' (ns/day) as a system with 32 6-core replicas. this means that we currently cannot make any assumptions about different coordination overhead for different number of replicas based on the troughput observed.